+++ CHANGELOG.md
@@ -4,7 +4,12 @@ This CHANGELOG file should help that the library becomes a standardized open sou
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2019-09-01
+## [Unreleased] - 2019-09-10
+
+### Added
+- Low Level funtion LogoClient::ReadBlock
+- Security function LogoClient::SetSessionPassword
+- Security function LogoClient::ClearSessionPassword
 
 ### Changed
 - Protection Level in GetProtection